# api/openapi.yaml
openapi: 3.0.3
info:
  title: SurveyFlow API
  description: SurveyFlow backend API for creating, distributing, and analyzing surveys.
  version: 1.0.0
  contact:
    name: SurveyFlow
    email: api@surveyflow.io
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/register:
    post:
      operationId: register
      tags: [Auth]
      summary: Register
      description: Create a new user account and organization.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Login
      description: Authenticate with email and password.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/login/2fa:
    post:
      operationId: login2FA
      tags: [Auth]
      summary: Login with 2FA
      description: Verify a two-factor authentication code.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
      responses:
        "200":
          description: 2FA verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [Auth]
      summary: Refresh token
      description: Exchange a refresh token for new access and refresh tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [Auth]
      summary: Forgot password
      description: Request a password reset email. Always returns 200.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Reset email sent (always)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/v1/auth/reset-password:
    post:
      operationId: resetPassword
      tags: [Auth]
      summary: Reset password
      description: Reset password using a token from the forgot-password email.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
                  minLength: 8
                  maxLength: 128
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/me:
    get:
      operationId: getMe
      tags: [Users]
      summary: Get current user
      description: Return the authenticated user's profile.
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateMe
      tags: [Users]
      summary: Update profile
      description: Update the current user's name or avatar.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /api/v1/me/password:
    put:
      operationId: changePassword
      tags: [Users]
      summary: Change password
      description: Change the current user's password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/organizations:
    post:
      operationId: createOrganization
      tags: [Organizations]
      summary: Create organization
      description: Create a new organization.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrgRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "409":
          description: Slug already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/organizations/me:
    get:
      operationId: getMyOrganization
      tags: [Organizations]
      summary: Get my organization
      description: Get the current user's organization.
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    put:
      operationId: updateMyOrganization
      tags: [Organizations]
      summary: Update my organization
      description: Update the current user's organization.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrgRequest"
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /api/v1/organizations/{orgId}:
    get:
      operationId: getOrganization
      tags: [Organizations]
      summary: Get organization
      description: Get an organization by ID (user must be a member).
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateOrganization
      tags: [Organizations]
      summary: Update organization
      description: Update an organization by ID.
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrgRequest"
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    delete:
      operationId: deleteOrganization
      tags: [Organizations]
      summary: Delete organization
      description: Delete an organization.
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "501":
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StubError"
  /api/v1/organizations/{orgId}/members:
    get:
      operationId: listMembers
      tags: [Organizations]
      summary: List members
      description: List all members of an organization.
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrgMember"
  /api/v1/organizations/{orgId}/members/invite:
    post:
      operationId: inviteMember
      tags: [Organizations]
      summary: Invite member
      description: Invite a new member to the organization.
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
        "409":
          description: Already a member or invitation pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/organizations/{orgId}/members/{memberId}/role:
    patch:
      operationId: updateMemberRole
      tags: [Organizations]
      summary: Update member role
      description: Change a member's role in the organization.
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/MemberId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [owner, admin, member, viewer]
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Cannot change owner role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/organizations/{orgId}/members/{memberId}:
    delete:
      operationId: removeMember
      tags: [Organizations]
      summary: Remove member
      description: Remove a member from the organization.
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/MemberId"
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Cannot remove owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/organizations/{orgId}/invitations:
    get:
      operationId: listInvitations
      tags: [Organizations]
      summary: List invitations
      description: List pending invitations for an organization.
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          description: Invitation list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invitation"
  /api/v1/invitations/{token}/accept:
    post:
      operationId: acceptInvitation
      tags: [Organizations]
      summary: Accept invitation
      description: Accept an organization invitation.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgMembership"
  /api/v1/invitations/{token}/decline:
    post:
      operationId: declineInvitation
      tags: [Organizations]
      summary: Decline invitation
      description: Decline an organization invitation.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invitation declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /api/v1/surveys:
    post:
      operationId: createSurvey
      tags: [Surveys]
      summary: Create survey
      description: Create a new survey.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyRequest"
      responses:
        "201":
          description: Survey created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listSurveys
      tags: [Surveys]
      summary: List surveys
      description: List surveys for the current organization with filtering and pagination.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, closed]
        - name: search
          in: query
          schema:
            type: string
        - name: sort_by
          in: query
          schema:
            type: string
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Survey list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"
  /api/v1/surveys/{surveyId}:
    get:
      operationId: getSurvey
      tags: [Surveys]
      summary: Get survey
      description: Get a survey by ID.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: Survey details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "404":
          description: Survey not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Wrong organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateSurvey
      tags: [Surveys]
      summary: Update survey
      description: Update a survey (partial update).
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSurveyRequest"
      responses:
        "200":
          description: Survey updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
    delete:
      operationId: deleteSurvey
      tags: [Surveys]
      summary: Delete survey
      description: Soft-delete a survey.
      parameters:
        $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: Survey deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/v1/surveys/{surveyId}/publish:
    post:
      operationId: publishSurvey
      tags: [Surveys]
      summary: Publish survey
      description: Publish a survey so it can accept responses.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: Survey published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "400":
          description: Survey has no questions or not in draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/surveys/{surveyId}/close:
    post:
      operationId: closeSurvey
      tags: [Surveys]
      summary: Close survey
      description: Close a published survey to stop accepting responses.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "200":
          description: Survey closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
  /api/v1/surveys/{surveyId}/duplicate:
    post:
      operationId: duplicateSurvey
      tags: [Surveys]
      summary: Duplicate survey
      description: Create a copy of a survey.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
      responses:
        "201":
          description: Survey duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
  /api/v1/surveys/{surveyId}/qr-code:
    get:
      operationId: getQRCode
      tags: [Distribution]
      summary: Get QR code
      description: Generate a QR code for a survey.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
        - name: size
          in: query
          schema:
            type: integer
        - name: error_level
          in: query
          schema:
            type: string
            enum: [L, M, Q, H]
      responses:
        "200":
          description: QR code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QRCodeResponse"
  /api/v1/surveys/{surveyId}/embed-code:
    get:
      operationId: getEmbedCode
      tags: [Distribution]
      summary: Get embed code
      description: Get HTML/JS embed code for a survey.
      parameters:
        - $ref: "#/components/parameters/SurveyId"
        - name: mode
          in: query
          schema:
            type: string
            enum: [popup, embedded, fullpage]
        - name: trigger
          in: query
          schema:
            type: string
            enum: [auto, button, custom]
        - name: width
          in: query
          schema:
            type: string
        - name: height
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Embed code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedCodeResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: invalid request body
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 255
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
    TokenResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            expires_at:
              type: string
              format: date-time
    AuthResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            tokens:
              type: object
              properties:
                access_token:
                  type: string
                refresh_token:
                  type: string
                expires_at:
                  type: string
                  format: date-time
...